Putting embedded opentype (iefix) back up to top of src list
This is so embedded opentype (iefix) is second to .eot src line in font-faces. The reason for this is internet explorer users will get their fonts seen first, requested, downloaded, and used as normal. For every other browser, it will go passed the .eot and .eot?iefix as it doesn't understand the file, and hit .woff2 (which is the sole preloaded web font file).

This ensures IE users get their fonts ASAP, and every other browser gets their .woff2 web fonts ASAP due to .woff2 web fonts being preloaded.

Note: This is the default src ordering of the google-webfonts-helper tool.