Improved security of lazyload-css.js and lazyload-css-min.js
For lazyload-css.js, I've made two explicit mini-functions for escaping potentially unsafe HTML and unescaping safe HTML. In the former mini-function, the unescaped string is checked with strict equality against any of the 3 vetted strings (also escaped). If it is passed, the output will then be unescaped to then move onto the next step of the function.

For lazyload-css-min.js, I've used the advanced optimisation for more savings.