Made JMP more concise
Made each sentence and paragraph of Introduction--Section 5 more concise and simple in terms of writing style, word choice, and flow. Jargon and technical terminology remained the same when it was needed. Otherwise, it was removed.

This so far resulted in a reduction of the overall PDF length by 10 pages.

I will continue these efforts for the other sections and conclusion tomorrow.